+++ NEWS.rst
@@ -5,6 +5,9 @@ image_calc (:numref:`image_calc`):
   * For a single input image, the option ``-c`` (calculation operation) defaults
     to ``var_0`` (identity operation).
   * The default output data type is now ``float32`` (previously ``float64``).
+  * Added the options ``--stretch`` and ``--percentile-range`` to stretch an
+    image and save it with 8-bit pixels, for visualization purposes
+    (:numref:`image_calc_stretch`).
 
 RELEASE 3.6.0, December 26, 2025
 --------------------------------